pyproject.toml
[project]
name = "fcs-ql-parser"
version = "1.2.0"
description = "FCS-QL (CLARIN-FCS Core 2.0 Query Language) Grammar and Parser"
readme = "README.md"
license = "MIT"
authors = [
    { name = "Erik Körner", email = "koerner@saw-leipzig.de" }
]
keywords = [
    "FCS",
    "FCS-QL",
    "CLARIN",
    "Query Parser"
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Text Processing :: General",
    "Topic :: Utilities",
]
requires-python = ">=3.10"
dependencies = [
    "antlr4-python3-runtime>=4.13.2",
]

[project.urls]
Source = "https://github.com/Querela/fcs-ql-python/"
Documentation = "https://fcs-ql-python.readthedocs.io/"
Issues = "https://github.com/Querela/fcs-ql-python/issues"

[project.optional-dependencies]
antlr = [
    "antlr4-tools>=0.2.2",
]
test = [
    "pytest>=9.0.2",
    "pytest-clarity>=1.0.1",
    "pytest-cov>=7.0.0",
    "pytest-randomly>=4.0.1",
]
style = [
    "black>=26.1.0",
    "flake8>=7.3.0",
    "isort>=8.0.0",
    "mypy>=1.19.1",
    "types-antlr4-python3-runtime>=4.13.0.20251118",
]
build = [
    "twine>=6.2.0",
]
docs = [
    "myst-parser>=4.0.1",
    "sphinx>=8.1.3",
    "sphinx-rtd-theme>=3.1.0",
]

[build-system]
requires = ["uv_build>=0.10.2,<0.11.0"]
build-backend = "uv_build"

[tool.uv.build-backend]
module-name = "fcsql"
source-exclude = [".antlr"]

[tool.black]
line-length = 120
extend-exclude = '''
(
  (FCS(Lexer|Parser(Listener|Visitor)?)).py
)
'''

[tool.isort]
profile = "black"
force_single_line = true
line_length = 120
known_first_party = "fcsql"
default_section = "THIRDPARTY"
forced_separate = "test_fcsql"
skip = [
    ".venv",
    "dist",
    "src/fcsql/FCSParser.py",
    "src/fcsql/FCSLexer.py",
    "src/fcsql/FCSParserListener.py",
    "src/fcsql/FCSParserVisitor.py",
]
skip_gitignore = true

# ignore all auto-generated antlr4 files,
# it needs a lot of work to quieten all warnings
[tool.mypy]
exclude = [
    'FCSLexer\.py$',
    'FCSParser\.py$',
    'FCSParserListener\.py$',
    'FCSParserVisitor\.py$',
    '^docs/',
]

[[tool.mypy.overrides]]
module = [
    "antlr4",
    "antlr4.error",
    "antlr4.error.ErrorListener",
]
ignore_missing_imports = true
